+++ Operations/CleanUpContainers.cs
@@ -1,14 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using McMaster.Extensions.CommandLineUtils;
-using Azure.Storage.Blobs;
+﻿using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
+using System;
+using System.Collections.Generic;
 using System.Globalization;
-using Microsoft.Extensions.Hosting;
-using static System.Runtime.InteropServices.JavaScript.JSType;
-using System.ComponentModel;
-using System.Reflection;
+using System.Linq;
 
 namespace LogicAppAdvancedTool.Operations
 {
@@ -22,19 +17,19 @@ namespace LogicAppAdvancedTool.Operations
             List<string> containerPrefixs = new List<string>();
             if (string.IsNullOrEmpty(workflowName))
             {
-                containerPrefixs.Add(CommonOperations.GenerateLogicAppPrefix());
+                containerPrefixs.Add(StoragePrefixGenerator.GenerateLogicAppPrefix());
             }
             else
             {
                 List<string> flowIDs = CommonOperations.ListFlowIDsByName(workflowName);
                 foreach (string flowID in flowIDs)
                 {
-                    containerPrefixs.Add(CommonOperations.GenerateWorkflowTablePrefixByFlowID(flowID));
+                    containerPrefixs.Add(StoragePrefixGenerator.GenerateWorkflowTablePrefixByFlowID(flowID));
                 }
             }
 
             List<string> matchedContainers = new List<string>();
-            BlobServiceClient client = new BlobServiceClient(AppSettings.ConnectionString);
+            BlobServiceClient client = StorageClientCreator.GenerateBlobServiceClient();
 
             foreach (string prefix in containerPrefixs)
             {