+++ Tools/Debugging.cs
@@ -1,6 +1,16 @@
-﻿using LogicAppAdvancedTool.Structures;
+﻿using Azure.Core;
+using Azure.Data.Tables;
+using Azure.Identity;
+using LogicAppAdvancedTool.Operations;
+using LogicAppAdvancedTool.Structures;
 using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
+using System.Linq;
+using System.Runtime.CompilerServices;
+using System.Security.Cryptography;
 
 namespace LogicAppAdvancedTool
 {